+++ src/main/java/core/impl/SearchWorkerImpl.java
@@ -3,8 +3,7 @@ package core.impl;
 
 import core.constants.CoreConstants;
 import core.contracts.*;
-import core.nnue.NnueManager;
-import core.nnue.NnueState;
+import core.records.NNUEState;
 import core.records.SearchInfo;
 import core.records.SearchResult;
 import core.records.SearchSpec;
@@ -19,8 +18,8 @@ import java.util.concurrent.locks.ReentrantLock;
 import static core.constants.CoreConstants.*;
 import static core.contracts.PositionFactory.*;
 
-public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
-    private final LazySmpWorkerPoolImpl pool;
+public final class SearchWorkerImpl implements Runnable, SearchWorker {
+    private final WorkerPoolImpl pool;
     final boolean isMainThread;
 
     /* ── threading primitives ─────────── */
@@ -39,7 +38,10 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
     private InfoHandler ih;
     private TranspositionTable tt;
     private MoveOrderer moveOrderer;
-    private final NnueState nnueState = new NnueState();
+
+    /* ── NNUE ────────── */
+    private final NNUEState nnueState = new NNUEState();
+    private NNUE nnue;
 
     private int lastScore;
     private boolean mateScore;
@@ -95,7 +97,7 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
         }
     }
 
-    public LazySmpSearchWorkerImpl(boolean isMainThread, LazySmpWorkerPoolImpl pool) {
+    public SearchWorkerImpl(boolean isMainThread, WorkerPoolImpl pool) {
         this.isMainThread = isMainThread;
         this.pool = pool;
         for (int i = 0; i < frames.length; ++i) {
@@ -168,7 +170,7 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
         for (int[] row : history) {
             Arrays.fill(row, 0);
         }
-        NnueManager.refreshAccumulator(nnueState, rootBoard);
+        NNUEImpl.refreshAccumulator(nnueState, rootBoard);
         // Change: Pass history to move orderer
         this.moveOrderer = new MoveOrdererImpl(history);
 
@@ -308,7 +310,7 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
                     return 0;
                 }
             }
-            if (ply >= MAX_PLY) return NnueManager.evaluateFromAccumulator(nnueState, PositionFactory.whiteToMove(bb[META]));
+            if (ply >= MAX_PLY) return nnue.evaluateFromAccumulator(nnueState, PositionFactory.whiteToMove(bb[META]));
         }
 
         boolean isPvNode = (beta - alpha) > 1;
@@ -355,7 +357,7 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
 
         // If not from TT (or if SCORE_NONE was stored), calculate it.
         if (staticEval == Integer.MIN_VALUE) {
-            staticEval = NnueManager.evaluateFromAccumulator(nnueState, PositionFactory.whiteToMove(bb[META]));
+            staticEval = nnue.evaluateFromAccumulator(nnueState, PositionFactory.whiteToMove(bb[META]));
         }
 
         // This prunes branches where the static evaluation is so high that it's
@@ -448,7 +450,7 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
             // --- Futility Pruning (Enhanced with History and Killers) ---
             if (!isPvNode && !inCheck && bestScore > -SCORE_MATE_IN_MAX_PLY && !isTactical) {
                 if (staticEval == Integer.MIN_VALUE) { // Calculate static eval if not already done
-                    staticEval = NnueManager.evaluateFromAccumulator(nnueState, PositionFactory.whiteToMove(bb[META]));
+                    staticEval = nnue.evaluateFromAccumulator(nnueState, PositionFactory.whiteToMove(bb[META]));
                 }
 
                 // Use a reduced depth similar to LMR to determine the margin
@@ -464,7 +466,7 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
 
             if (!pf.makeMoveInPlace(bb, mv, mg)) continue;
             legalMovesFound++;
-            updateNnueAccumulator(moverPiece, capturedPiece, mv);
+            nnue.updateNnueAccumulator(nnueState, moverPiece, capturedPiece, mv);
 
             int score;
             if (i == 0) {
@@ -487,7 +489,7 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
             }
 
             pf.undoMoveInPlace(bb);
-            undoNnueAccumulatorUpdate(moverPiece, capturedPiece, mv);
+            nnue.undoNnueAccumulatorUpdate(nnueState, moverPiece, capturedPiece, mv);
             if (pool.isStopped()) return 0;
 
             if (ply == 0) {
@@ -547,7 +549,7 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
             }
         }
 
-        if (ply >= MAX_PLY) return NnueManager.evaluateFromAccumulator(nnueState, PositionFactory.whiteToMove(bb[META]));
+        if (ply >= MAX_PLY) return nnue.evaluateFromAccumulator(nnueState, PositionFactory.whiteToMove(bb[META]));
 
         nodes++;
 
@@ -574,12 +576,12 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
                 legalMovesFound++;
 
                 // --- Update NNUE state AFTER making the move ---
-                updateNnueAccumulator(moverPiece, capturedPiece, mv);
+                nnue.updateNnueAccumulator(nnueState, moverPiece, capturedPiece, mv);
 
                 int score = -quiescence(bb, -beta, -alpha, ply + 1);
 
                 // --- Undo NNUE state BEFORE undoing the move ---
-                undoNnueAccumulatorUpdate(moverPiece, capturedPiece, mv);
+                nnue.undoNnueAccumulatorUpdate(nnueState, moverPiece, capturedPiece, mv);
                 pf.undoMoveInPlace(bb);
 
                 if (pool.isStopped()) return 0;
@@ -600,7 +602,7 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
         // Logic Path 2: The king is NOT in check.
         else {
             // Use the NNUE accumulator for the stand-pat evaluation.
-            int standPat = NnueManager.evaluateFromAccumulator(nnueState, PositionFactory.whiteToMove(bb[META]));
+            int standPat = nnue.evaluateFromAccumulator(nnueState, PositionFactory.whiteToMove(bb[META]));
             if (standPat >= beta) return beta;
             if (standPat > alpha) alpha = standPat;
 
@@ -624,12 +626,12 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
                 if (!pf.makeMoveInPlace(bb, mv, mg)) continue;
 
                 // --- Update NNUE state AFTER making the move ---
-                updateNnueAccumulator(moverPiece, capturedPiece, mv);
+                nnue.updateNnueAccumulator(nnueState, moverPiece, capturedPiece, mv);
 
                 int score = -quiescence(bb, -beta, -alpha, ply + 1);
 
                 // --- Undo NNUE state BEFORE undoing the move ---
-                undoNnueAccumulatorUpdate(moverPiece, capturedPiece, mv);
+                nnue.undoNnueAccumulatorUpdate(nnueState, moverPiece, capturedPiece, mv);
                 pf.undoMoveInPlace(bb);
 
                 if (pool.isStopped()) return 0;
@@ -708,68 +710,8 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
         return -1; // No capture
     }
 
-    /**
-     * Updates the NNUE accumulator based on a move, now correctly handling castling.
-     */
-    private void updateNnueAccumulator(int moverPiece, int capturedPiece, int move) {
-        int from = (move >>> 6) & 0x3F;
-        int to = move & 0x3F;
-        int moveType = (move >>> 14) & 0x3;
-
-        if (capturedPiece != -1) {
-            int capturedSquare = (moveType == 2) ? (to + (moverPiece < 6 ? -8 : 8)) : to;
-            NnueManager.removePiece(nnueState, capturedPiece, capturedSquare);
-        }
-
-        if (moveType == 1) { // Promotion
-            int promotedToPiece = (moverPiece < 6 ? WN : BN) + ((move >>> 12) & 0x3);
-            NnueManager.removePiece(nnueState, moverPiece, from);
-            NnueManager.addPiece(nnueState, promotedToPiece, to);
-        } else if (moveType == 3) { // Castle
-            int rook = moverPiece < 6 ? WR : BR;
-            switch(to) {
-                case 6: NnueManager.updateCastle(nnueState, WK, rook, 4, 6, 7, 5); break; // White O-O
-                case 2: NnueManager.updateCastle(nnueState, WK, rook, 4, 2, 0, 3); break; // White O-O-O
-                case 62: NnueManager.updateCastle(nnueState, BK, rook, 60, 62, 63, 61); break; // Black O-O
-                case 58: NnueManager.updateCastle(nnueState, BK, rook, 60, 58, 56, 59); break; // Black O-O-O
-            }
-        } else { // Normal move
-            NnueManager.updateAccumulator(nnueState, moverPiece, from, to);
-        }
-    }
-
-    /**
-     * Undoes a move's effect on the NNUE accumulator, now correctly handling castling.
-     */
-    private void undoNnueAccumulatorUpdate(int moverPiece, int capturedPiece, int move) {
-        int from = (move >>> 6) & 0x3F;
-        int to = move & 0x3F;
-        int moveType = (move >>> 14) & 0x3;
-
-        if (moveType == 1) { // Promotion
-            int promotedToPiece = (moverPiece < 6 ? WN : BN) + ((move >>> 12) & 0x3);
-            NnueManager.removePiece(nnueState, promotedToPiece, to);
-            NnueManager.addPiece(nnueState, moverPiece, from);
-        } else if (moveType == 3) { // Castle
-            int rook = moverPiece < 6 ? WR : BR;
-            switch(to) {
-                case 6: NnueManager.updateCastle(nnueState, WK, rook, 6, 4, 5, 7); break; // Undo White O-O
-                case 2: NnueManager.updateCastle(nnueState, WK, rook, 2, 4, 3, 0); break; // Undo White O-O-O
-                case 62: NnueManager.updateCastle(nnueState, BK, rook, 62, 60, 61, 63); break; // Undo Black O-O
-                case 58: NnueManager.updateCastle(nnueState, BK, rook, 58, 60, 59, 56); break; // Undo Black O-O-O
-            }
-        } else { // Normal move
-            NnueManager.updateAccumulator(nnueState, moverPiece, to, from);
-        }
-
-        if (capturedPiece != -1) {
-            int capturedSquare = (moveType == 2) ? (to + (moverPiece < 6 ? -8 : 8)) : to;
-            NnueManager.addPiece(nnueState, capturedPiece, capturedSquare);
-        }
-    }
-
     @Override
-    public void prepareForSearch(long[] root, SearchSpec s, PositionFactory p, MoveGenerator m, Evaluator e, TranspositionTable t, TimeManager timeMgr) {
+    public void prepareForSearch(long[] root, SearchSpec s, PositionFactory p, MoveGenerator m, NNUE nnue, TranspositionTable t, TimeManager timeMgr) {
         this.rootBoard = root.clone();
         this.spec = s;
         this.pf = p;
@@ -777,6 +719,7 @@ public final class LazySmpSearchWorkerImpl implements Runnable, SearchWorker {
         this.tt = t;
         this.tm = timeMgr;
         this.gameHistory = s.history();
+        this.nnue = nnue;
     }
 
     public void startWorkerSearch() {
